Restrict dropping non-empty schema

+++ presto-main/src/main/java/io/prestosql/execution/DropSchemaTask.java
@@ -16,6 +16,7 @@ package io.prestosql.execution;
 import com.google.common.util.concurrent.ListenableFuture;
 import io.prestosql.Session;
 import io.prestosql.metadata.Metadata;
+import io.prestosql.metadata.QualifiedTablePrefix;
 import io.prestosql.security.AccessControl;
 import io.prestosql.spi.PrestoException;
 import io.prestosql.spi.connector.CatalogSchemaName;
@@ -29,6 +30,7 @@ import java.util.Optional;
 import static com.google.common.util.concurrent.Futures.immediateFuture;
 import static io.prestosql.metadata.MetadataUtil.createCatalogSchemaName;
 import static io.prestosql.spi.StandardErrorCode.NOT_SUPPORTED;
+import static io.prestosql.spi.StandardErrorCode.SCHEMA_NOT_EMPTY;
 import static io.prestosql.spi.StandardErrorCode.SCHEMA_NOT_FOUND;
 import static io.prestosql.sql.analyzer.SemanticExceptions.semanticException;
 
@@ -64,10 +66,30 @@ public class DropSchemaTask
             return immediateFuture(null);
         }
 
+        if (!isSchemaEmpty(session, schema, metadata)) {
+            throw semanticException(SCHEMA_NOT_EMPTY, statement, "Cannot drop non-empty schema '%s'", schema.getSchemaName());
+        }
+
         accessControl.checkCanDropSchema(session.toSecurityContext(), schema);
 
         metadata.dropSchema(session, schema);
 
         return immediateFuture(null);
+    }
+
+    private static boolean isSchemaEmpty(Session session, CatalogSchemaName schema, Metadata metadata)
+    {
+        QualifiedTablePrefix tablePrefix = new QualifiedTablePrefix(schema.getCatalogName(), schema.getSchemaName());
+
+        // These are best efforts checks that don't provide any guarantees against concurrent DDL operations
+        if (!metadata.listTables(session, tablePrefix).isEmpty()) {
+            return false;
+        }
+
+        if (!metadata.listViews(session, tablePrefix).isEmpty()) {
+            return false;
+        }
+
+        return true;
     }
 }
